move navbar to dedicated file
